fix(indexer): gate slow-mount-magic test to Linux, unbreak macOS build
TestIsSlowMountFSType exercises isSlowMountFSType, which only exists in
slow_mount_linux.go (//go:build linux). The test file had no build tag,
so it compiled on every platform, and on macOS/Windows (only
slow_mount_other.go compiled, defining slowWatchMount but not
isSlowMountFSType) the package failed to build:

    internal/indexer/slow_mount_test.go:40:14: undefined: isSlowMountFSType

Split the file: TestIsSlowMountFSType stays behind //go:build linux
alongside the function it pins; TestSlowWatchMount_NormalMountNotDegraded
(exercises the cross-platform slowWatchMount) moves to a new untagged
file so it keeps running on macOS/Windows too.
